# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Intercultural Experience Platform API",
        "description": "API completa para la plataforma de intercambios culturales. Incluye autenticación, gestión de programas, aplicaciones, sistema de puntos y más.",
        "contact": {
            "name": "IE Development Team",
            "email": "developers@ie.org.py"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://ie.org.py/terms"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ie.org.py/api",
            "description": "Servidor de Producción"
        },
        {
            "url": "https://staging.ie.org.py/api",
            "description": "Servidor de Staging"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
...
        "/english-evaluations": {
            "get": {
                "tags": [
                    "English Tests"
                ],
                "summary": "Listar evaluaciones de inglés del usuario",
                "description": "Obtiene todas las evaluaciones de inglés del usuario autenticado, ordenadas por fecha descendente",
                "operationId": "5128f8cd8ce61e4198f189b6d8f9e87d",
                "responses": {
                    "200": {
                        "description": "Lista de evaluaciones obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "score": {
                                                        "type": "integer",
                                                        "example": 75
                                                    },
                                                    "cefr_level": {
                                                        "type": "string",
                                                        "example": "B2"
                                                    },
                                                    "ef_set_id": {
                                                        "type": "string",
                                                        "example": "EF123456"
                                                    },
                                                    "evaluated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-10-20T14:30:00Z"
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Excelente comprensión oral"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "remaining_attempts": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "can_attempt": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "English Tests"
                ],
                "summary": "Registrar nueva evaluación de inglés",
                "description": "Registra una nueva evaluación de inglés. Máximo 3 intentos por usuario. El nivel CEFR se calcula automáticamente según el puntaje.",
                "operationId": "dce8b7507397a8112313f151ddef35d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "score"
                                ],
                                "properties": {
                                    "score": {
                                        "description": "Puntaje obtenido (0-100)",
                                        "type": "integer",
                                        "maximum": 100,
                                        "minimum": 0,
                                        "example": 75
                                    },
                                    "ef_set_id": {
                                        "description": "ID del test EF SET (opcional)",
                                        "type": "string",
                                        "example": "EF123456"
                                    },
                                    "notes": {
                                        "description": "Notas adicionales (opcional)",
                                        "type": "string",
                                        "example": "Test realizado en línea"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evaluación registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Evaluación registrada exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "score": {
                                                    "type": "integer",
                                                    "example": 75
                                                },
                                                "cefr_level": {
                                                    "type": "string",
                                                    "example": "B2"
                                                },
                                                "classification": {
                                                    "type": "string",
                                                    "example": "Upper Intermediate"
                                                },
                                                "attempt_number": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "remaining_attempts": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Límite de intentos alcanzado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}